# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Law Firm Ledger API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Client {
  id: number;
  name: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface ClientInput {
  /** @minLength 1 */
  name: string;
  email?: string;
  phone?: string;
  address?: string;
  notes?: string;
}

export interface ClientUpdate {
  /** @minLength 1 */
  name?: string;
  email?: string;
  phone?: string;
  address?: string;
  notes?: string;
}

export interface Case {
  id: number;
  clientId: number;
  /** @nullable */
  clientName?: string | null;
  caseNo: string;
  date: string;
  /**
     * @minimum 1
     * @maximum 4
     */
  stage: number;
  /** @nullable */
  tmNumber?: string | null;
  /** @nullable */
  detail?: string | null;
  due: number;
  received: number;
  balance: number;
  createdAt: string;
}

export interface CaseInput {
  clientId: number;
  /** @minLength 1 */
  caseNo: string;
  date: string;
  /**
     * @minimum 1
     * @maximum 4
     */
  stage: number;
  tmNumber?: string;
  detail?: string;
  due: number;
  received: number;
}

export interface CaseUpdate {
  clientId?: number;
  caseNo?: string;
  date?: string;
  /**
     * @minimum 1
     * @maximum 4
     */
  stage?: number;
  tmNumber?: string;
  detail?: string;
  due?: number;
  received?: number;
}

export interface StageCount {
  stage: number;
  count: number;
}

export interface ReportSummary {
  totalClients: number;
  totalCases: number;
  totalDue: number;
  totalReceived: number;
  totalBalance: number;
  casesByStage: StageCount[];
}

export interface ClientReport {
  client: Client;
  cases: Case[];
  totalDue: number;
  totalReceived: number;
  totalBalance: number;
}

export type ListCasesParams = {
/**
 * @nullable
 */
clientId?: number | null;
/**
 * @nullable
 */
stage?: number | null;
};