# lib/api-zod/src/generated/types/case.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Law Firm Ledger API
 * OpenAPI spec version: 0.1.0
 */

export interface Case {
  id: number;
  clientId: number;
  /** @nullable */
  clientName?: string | null;
  caseNo: string;
  date: string;
  /**
     * @minimum 1
     * @maximum 4
     */
  stage: number;
  /** @nullable */
  tmNumber?: string | null;
  /** @nullable */
  detail?: string | null;
  due: number;
  received: number;
  balance: number;
  createdAt: string;
}