# lib/api-zod/src/generated/types/caseInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Law Firm Ledger API
 * OpenAPI spec version: 0.1.0
 */

export interface CaseInput {
  clientId: number;
  /** @minLength 1 */
  caseNo: string;
  date: string;
  /**
     * @minimum 1
     * @maximum 4
     */
  stage: number;
  tmNumber?: string;
  detail?: string;
  due: number;
  received: number;
}