# lib/api-zod/src/generated/types/client.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Law Firm Ledger API
 * OpenAPI spec version: 0.1.0
 */

export interface Client {
  id: number;
  name: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}