# lib/api-zod/src/generated/types/clientReport.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Law Firm Ledger API
 * OpenAPI spec version: 0.1.0
 */
import type { Case } from './case';
import type { Client } from './client';

export interface ClientReport {
  client: Client;
  cases: Case[];
  totalDue: number;
  totalReceived: number;
  totalBalance: number;
}