# lib/api-zod/src/generated/types/reportSummary.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Law Firm Ledger API
 * OpenAPI spec version: 0.1.0
 */
import type { StageCount } from './stageCount';

export interface ReportSummary {
  totalClients: number;
  totalCases: number;
  totalDue: number;
  totalReceived: number;
  totalBalance: number;
  casesByStage: StageCount[];
}